집에서 다시 커밋

+++ Trip/src/main/java/org/Trip/domain/ImageFileVO.java
@@ -15,7 +15,7 @@ public class ImageFileVO {
 	private String sub4;
 	private Long fno;
 	private Long tno;
-	private int localId;
+	private Long localId;
 	private Date regdate;
 	private Date update;
